Container version update
* Removing microrelease version numbers from image tags to pick up
  latest minors
* Bumping RHSSO 7.4 to 7.5
* Bumping RHSCL PostgreSQL to 13
* Bumping Oracle XE to 21
