+++ pom.xml
@@ -200,19 +200,19 @@
                                 <!-- Services used in test suite -->
                                 <keycloak.image>quay.io/keycloak/keycloak:15.0.1</keycloak.image>
                                 <rhsso.73.image>registry.access.redhat.com/redhat-sso-7/sso73-openshift</rhsso.73.image>
-                                <postgresql.13.image>quay.io/bitnami/postgresql:13.4.0</postgresql.13.image>
+                                <postgresql.13.image>quay.io/bitnami/postgresql:13.5.0</postgresql.13.image>
                                 <mysql.57.image>quay.io/bitnami/mysql:5.7.32</mysql.57.image>
                                 <mysql.80.image>registry.access.redhat.com/rhscl/mysql-80-rhel7</mysql.80.image>
-                                <mariadb.10.image>quay.io/quarkusqeteam/mariadb:10.6.4</mariadb.10.image>
+                                <mariadb.10.image>quay.io/bitnami/mariadb:10.6</mariadb.10.image>
                                 <mariadb.102.image>registry.access.redhat.com/rhscl/mariadb-102-rhel7</mariadb.102.image>
-                                <mssql.image>mcr.microsoft.com/mssql/rhel/server</mssql.image>
-                                <oracle.image>gvenzl/oracle-xe:18.4.0-slim</oracle.image>
+                                <mssql.image>mcr.microsoft.com/mssql/rhel/server:2019-latest</mssql.image>
+                                <oracle.image>gvenzl/oracle-xe:21-slim</oracle.image>
                                 <db2.image>quay.io/quarkusqeteam/db2:11.5.5.0</db2.image>
-                                <mongodb.image>quay.io/bitnami/mongodb:5.0.2</mongodb.image>
+                                <mongodb.image>quay.io/bitnami/mongodb:5.0</mongodb.image>
                                 <redis.image>quay.io/bitnami/redis:6.0</redis.image>
                                 <wiremock.image>quay.io/ocpmetal/wiremock</wiremock.image>
-                                <consul.image>quay.io/bitnami/consul:1.9.3</consul.image>
-                                <infinispan.image>infinispan/server:12.1</infinispan.image>
+                                <consul.image>quay.io/bitnami/consul:1.10</consul.image>
+                                <infinispan.image>quay.io/infinispan/server:12.1</infinispan.image>
                                 <infinispan.expected-log>Infinispan Server.*started in</infinispan.expected-log>
                                 <spring.cloud.server.image>quay.io/quarkusqeteam/spring-cloud-config-server:3.0</spring.cloud.server.image>
                             </systemPropertyVariables>
@@ -587,9 +587,9 @@
                             <systemPropertyVariables>
                                 <ts.redhat.registry.enabled>true</ts.redhat.registry.enabled>
                                 <!-- Product Services -->
-                                <rhsso.74.image>registry.redhat.io/rh-sso-7/sso74-openshift-rhel8</rhsso.74.image>
+                                <rhsso.75.image>registry.redhat.io/rh-sso-7/sso75-openshift-rhel8</rhsso.75.image>
                                 <postgresql.10.image>registry.redhat.io/rhscl/postgresql-10-rhel7</postgresql.10.image>
-                                <postgresql.12.image>registry.redhat.io/rhscl/postgresql-12-rhel7</postgresql.12.image>
+                                <postgresql.13.image>registry.redhat.io/rhscl/postgresql-13-rhel7</postgresql.13.image>
                                 <mariadb.103.image>registry.redhat.io/rhscl/mariadb-103-rhel7</mariadb.103.image>
                                 <amq-streams.image>registry.redhat.io/amq7/amq-streams-kafka-27-rhel7</amq-streams.image>
                                 <amq-streams.version>1.7.0</amq-streams.version>
