+++ security/keycloak-multitenant/src/test/java/io/quarkus/ts/security/keycloak/multitenant/OpenShiftRhSso75MultiTenantSecurityIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.security;
+package io.quarkus.ts.security.keycloak.multitenant;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
@@ -10,18 +10,17 @@ import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
 @EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftRhSso74Oauth2SecurityIT extends BaseOauth2SecurityIT {
+public class OpenShiftRhSso75MultiTenantSecurityIT extends BaseMultiTenantSecurityIT {
 
     static final int KEYCLOAK_PORT = 8080;
 
-    @Container(image = "${rhsso.74.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
+    @Container(image = "${rhsso.75.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
     static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
             .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
 
     @QuarkusApplication
     static RestService app = new RestService()
-            .withProperty("quarkus.oauth2.introspection-url",
-                    () -> keycloak.getRealmUrl() + "/protocol/openid-connect/token/introspect");
+            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl());
 
     @Override
     protected KeycloakService getKeycloak() {
