+++ security/keycloak-webapp/src/test/java/io/quarkus/ts/security/keycloak/webapp/OpenShiftRhSso75WebappSecurityIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.security.keycloak.authz;
+package io.quarkus.ts.security.keycloak.webapp;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
@@ -10,19 +10,18 @@ import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
 @EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftRhSso74AuthzSecurityIT extends BaseAuthzSecurityIT {
+public class OpenShiftRhSso75WebappSecurityIT extends BaseWebappSecurityIT {
 
     static final int KEYCLOAK_PORT = 8080;
 
-    @Container(image = "${rhsso.74.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
+    @Container(image = "${rhsso.75.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
     static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
             .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
 
     @QuarkusApplication
     static RestService app = new RestService()
             .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl())
-            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT)
-            .withProperty("quarkus.oidc.credentials.secret", CLIENT_SECRET_DEFAULT);
+            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT);
 
     @Override
     protected KeycloakService getKeycloak() {
