+++ security/keycloak/src/test/java/io/quarkus/ts/security/OpenShiftRhSso75OidcSecurityIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.security.keycloak.webapp;
+package io.quarkus.ts.security;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
@@ -10,26 +10,22 @@ import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
 @EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftRhSso74WebappSecurityIT extends BaseWebappSecurityIT {
+public class OpenShiftRhSso75OidcSecurityIT extends BaseOidcSecurityIT {
 
     static final int KEYCLOAK_PORT = 8080;
 
-    @Container(image = "${rhsso.74.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
+    @Container(image = "${rhsso.75.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
     static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
             .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
 
     @QuarkusApplication
     static RestService app = new RestService()
             .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl())
-            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT);
+            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT)
+            .withProperty("quarkus.oidc.credentials.secret", CLIENT_SECRET_DEFAULT);
 
     @Override
     protected KeycloakService getKeycloak() {
         return keycloak;
-    }
-
-    @Override
-    protected RestService getApp() {
-        return app;
     }
 }
