+++ sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/MariaDB10DatabaseIT.java
@@ -11,7 +11,7 @@ public class MariaDB10DatabaseIT extends AbstractSqlDatabaseIT {
 
     static final int MARIADB_PORT = 3306;
 
-    @Container(image = "${mariadb.10.image}", port = MARIADB_PORT, expectedLog = "MariaDB init process done. Ready for start up")
+    @Container(image = "${mariadb.10.image}", port = MARIADB_PORT, expectedLog = "/opt/bitnami/mariadb/sbin/mysqld: ready for connections.")
     static MariaDbService database = new MariaDbService();
 
     @QuarkusApplication
